+++ bench/src/benchmarks/check.bench.ts
@@ -1,15 +1,12 @@
-import type { SplitResult } from "@filego/js";
-import type { SplitResult as NodeSplitResult } from "@filego/node";
+import type { SplitResult } from "@filego/node";
 
-import { check } from "@filego/js";
-import { check as nodeCheck } from "@filego/node";
+import { check } from "@filego/node";
 import { bench, describe } from "vitest";
 
-import { cacheDir, preNodeSplit, preSplit } from "#/configs";
+import { cacheDir, preSplit } from "#/configs";
 
 describe("Check", async (): Promise<void> => {
     const splitted: SplitResult = await preSplit();
-    const nodeSplitted: NodeSplitResult = await preNodeSplit();
 
     bench("JS", async (): Promise<void> => {
         await check({
@@ -20,10 +17,10 @@ describe("Check", async (): Promise<void> => {
     });
 
     bench("Node", async (): Promise<void> => {
-        await nodeCheck({
+        await check({
             inDir: cacheDir,
-            fileSize: nodeSplitted.fileSize,
-            totalChunks: nodeSplitted.totalChunks,
+            fileSize: splitted.fileSize,
+            totalChunks: splitted.totalChunks,
         });
     });
 });