+++ bench/src/benchmarks/split.bench.ts
@@ -1,5 +1,4 @@
-import { split } from "@filego/js";
-import { split as nodeSplit } from "@filego/node";
+import { split } from "@filego/node";
 import { bench, describe } from "vitest";
 
 import { cacheDir, chunkSize, getFile, inFile } from "#/configs";
@@ -13,7 +12,7 @@ describe("Split", (): void => {
     });
 
     bench("Node", async (): Promise<void> => {
-        await nodeSplit({
+        await split({
             inFile,
             outDir: cacheDir,
             chunkSize,