+++ packages/js/src/functions/check.ts
@@ -1,32 +1,6 @@
-import type { CheckResult, FileChunk } from "@filego/shared";
+import type { CheckOptions, CheckResult, FileChunk } from "@filego/shared";
 
-import { isTypeOfFileChunks, sortFileChunks } from "@filego/shared";
-
-type _CheckOptions = {
-    /** Chunks to be checked in the `check` function. */
-    chunks: FileChunk[];
-    /**
-     * Size of the original file,
-     * which can be found as the output of the `split` function.
-     */
-    fileSize: number;
-    /**
-     * Total number of chunks in the original file,
-     * which can be found as the output of the `split` function.
-     */
-    totalChunks: number;
-};
-
-/** Options for the custom check function in the `check` function. */
-type CheckFunctionOptions = _CheckOptions;
-
-/** Options for the `check` function. */
-type CheckOptions = _CheckOptions & {
-    /** Custom check function for the `check` function. */
-    checkFunction?: (
-        options: CheckFunctionOptions,
-    ) => CheckResult | Promise<CheckResult>;
-};
+import { sortFileChunks } from "@filego/shared";
 
 /**
  * This function checks file integrity by verifying
@@ -49,22 +23,6 @@ type CheckOptions = _CheckOptions & {
 const check = async (options: CheckOptions): Promise<CheckResult> => {
     const { chunks, fileSize, totalChunks, checkFunction }: CheckOptions =
         options;
-
-    if (isTypeOfFileChunks(chunks) === false) {
-        throw new TypeError("chunks is not an Array of Chunk");
-    }
-
-    if (typeof fileSize !== "number") {
-        throw new TypeError("fileSize is not a number");
-    }
-
-    if (typeof totalChunks !== "number") {
-        throw new TypeError("totalChunks is not a number");
-    }
-
-    if (checkFunction && typeof checkFunction !== "function") {
-        throw new TypeError("checkFunction is not a function");
-    }
 
     // custom check function
     if (checkFunction) {
@@ -119,5 +77,4 @@ const check = async (options: CheckOptions): Promise<CheckResult> => {
     };
 };
 
-export type { CheckOptions, CheckFunctionOptions, CheckResult };
 export { check };