+++ packages/node/src/functions/split.ts
@@ -1,68 +1,16 @@
 import type { ReadStream } from "node:fs";
 
-type _SplitOptions = {
-    /** Input file to be splitted in the `split` function. */
-    inFile: string;
-    /** Output directory after splitted in the `split` function. */
-    outDir: string;
-    /** Size of each chunk in byte to be splitted. */
-    chunkSize: number;
-};
+import type {
+    SplitIoOptions,
+    SplitIoResult,
+    SplitOptions,
+    SplitResult,
+} from "@filego/shared";
 
-/** Options for the custom split function in the `split` function. */
-type SplitFunctionOptions = _SplitOptions;
+import { split as _split } from "@filego/js";
 
-/** Result of the `split` function. */
-type SplitResult = {
-    /** Size of the original file. */
-    fileSize: number;
-    /** The total number of chunks splitted from the original file. */
-    totalChunks: number;
-};
-
-/** Options for the `split` function. */
-type SplitOptions = _SplitOptions & {
-    /** Custom split function for `split` function. */
-    splitFunction?: (
-        options: SplitFunctionOptions,
-    ) => SplitResult | Promise<SplitResult>;
-};
-
-/**
- * This function splits file from a path to a directory directly.
- *
- * It will only return the `fileSize` and the `totalChunks` of the file.
- *
- * ### Example
- *
- * ```ts
- * import { split } from "@filego/node";
- *
- * await split({
- *     inFile: "/path/to/file.txt",
- *     outDir: "/path/to/dir",
- *     chunkSize: 2 * 1024 * 1024,
- * });
- * ```
- */
-const split = async (options: SplitOptions): Promise<SplitResult> => {
-    const { inFile, outDir, chunkSize, splitFunction }: SplitOptions = options;
-
-    if (typeof inFile !== "string") {
-        throw new TypeError("inFile is not a path in string");
-    }
-
-    if (typeof outDir !== "string") {
-        throw new TypeError("outDir is not a directory in string");
-    }
-
-    if (typeof chunkSize !== "number") {
-        throw new TypeError("chunkSize is not a number");
-    }
-
-    if (splitFunction && typeof splitFunction !== "function") {
-        throw new TypeError("splitFunction is not a function");
-    }
+const splitIo = async (options: SplitIoOptions): Promise<SplitIoResult> => {
+    const { inFile, outDir, splitFunction, chunkSize } = options;
 
     // custom
 
@@ -144,5 +92,51 @@ const split = async (options: SplitOptions): Promise<SplitResult> => {
     return { fileSize, totalChunks: index };
 };
 
-export type { SplitOptions, SplitResult, SplitFunctionOptions };
+/**
+ * This function splits file into chunks,
+ * and the chunks can be used to check and merge later.
+ *
+ * It will return the `chunks`, `fileSize` and `totalChunks`.
+ *
+ * ### Example
+ *
+ * ```ts
+ * import { split } from "@filego/node";
+ *
+ * const file: File | Blob | Uint8Array | string = "";
+ *
+ * await split({
+ *     file,
+ *     chunkSize: 2 * 1024 * 1024,
+ * });
+ * ```
+ */
+async function split(options: SplitOptions): Promise<SplitResult>;
+
+/**
+ * This function splits file from a path to a directory directly.
+ *
+ * It will only return the `fileSize` and the `totalChunks` of the file.
+ *
+ * ### Example
+ *
+ * ```ts
+ * import { split } from "@filego/node";
+ *
+ * await split({
+ *     inFile: "/path/to/file.txt",
+ *     outDir: "/path/to/dir",
+ *     chunkSize: 2 * 1024 * 1024,
+ * });
+ * ```
+ */
+async function split(options: SplitIoOptions): Promise<SplitIoResult>;
+
+async function split(
+    options: SplitOptions | SplitIoOptions,
+): Promise<SplitResult | SplitIoResult> {
+    if ("file" in options) return await _split(options);
+    return await splitIo(options);
+}
+
 export { split };